slack divided.
